destpaths.c: avoid signed/unsigned-comparison warning
* fish/destpaths.c (free_words): Change param type: s/int/size_t/.

+++ fish/destpaths.c
@@ -61,7 +61,7 @@ struct word {
 };
 
 static void
-free_words (struct word *words, int nr_words)
+free_words (struct word *words, size_t nr_words)
 {
   size_t i;
